Tried implementing stepCount for distance calculation

+++ RunLibrary/src/main/java/com/example/runlibrary/Run.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 class Run(
-    val pace: Float,
+    val pace: Double,
     val distance: Double,
     val duration: Long,
     val timestamp: Long
